requirements.txt
# FastAPI core
fastapi
uvicorn[standard]

# PostgreSQL DB + ORM
sqlalchemy
asyncpg


# Authentication / Security
python-jose[cryptography]
passlib[bcrypt]

# Caching / Redis
redis
aioredis

# Background Jobs / Async tasks
rq     # Redis Queue (simple background jobs)
rq-scheduler  # Optional for scheduled jobs

# Pydantic for data validation
pydantic

# Email Sending
aiosmtplib

# Utilities
python-dotenv  # Load .env
httpx         # Async HTTP requests (if needed)
loguru        # Advanced logging

# Optional: WebSockets for real-time notifications
websockets

# Optional: LRU Cache (pure Python)
cachetools